Remove timeout
A 50 ms timeout was causing a lot of false negatives for certain test cases. Looking at the blame, I believe that this line had been changed to a higher value but then was readded as 50ms via a copy and paste error. I have decided to remove the timeout entirerly and let ATF use the default timeout instead.